fun delete AlertDialog confirmation

+++ app/src/main/java/com/company/convidades/view/viewholder/GuestViewHolder.kt
@@ -1,9 +1,8 @@
 package com.company.convidades.view.viewholder
 
-import android.view.View
-import android.widget.TextView
+import android.content.DialogInterface
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
-import com.company.convidades.R
 import com.company.convidades.databinding.RowGuestBinding
 import com.company.convidades.model.GuestModel
 import com.company.convidades.view.listener.IGuestListener
@@ -19,7 +18,18 @@ class GuestViewHolder(private val bind: RowGuestBinding, private val listener: I
         }
 
         bind.textName.setOnLongClickListener {
-            listener.onDelete(guestModel.id)
+
+            AlertDialog.Builder(itemView.context)
+                .setTitle("Atenção!")
+                .setMessage("Tem certeza que deseja excluir?")
+                .setPositiveButton(
+                    "Sim"
+                ) { dialog, which ->
+                    listener.onDelete(guestModel.id)
+                }
+                .setNegativeButton("Não", null)
+                .create().show()
+
             true
         }
     }
